[level 2] Title: 2 x n 타일링, Time: 2.23 ms, Memory: 52.3 MB -BaekjoonHub

+++ 프로그래머스/2/12900. 2 x n 타일링/2 x n 타일링.java
@@ -3,13 +3,13 @@ class Solution {
         if (n == 1) return 1;
 		else if (n == 2) return 2;
 		else {
-			long[] fn = new long[n + 1];
+			int[] fn = new int[n + 1];
 			fn[1] = 1;
 			fn[2] = 2;
 			for (int i = 3; i <= n; i++) {
 				fn[i] = (fn[i - 2] + fn[i - 1]) % 1000000007;
 			}
-			return (int) fn[n];
+			return fn[n];
 		}
     }
 }
